Link changes, better compliance with the AGPL, favicon, and improved CSS
Seeing as I just figured out how to use github pages, I figured it'd be
easier to host this project on there as opposed to a woefully outdated
hosting service.

I noticed that there was some incompliance with the AGPL in my files, so I
added a comment that would bring compliance with the AGPL.

I figured that, since I was updating the site anyways, I might as well add
a favicon.

I noticed that I was not being very optimal with my CSS a few days ago, so
I made some changes in the CSS.